+++ iSpendLiveActivity/iSpendLiveActivityBundle.swift
@@ -1,0 +1,7 @@
+import WidgetKit
+import SwiftUI
+
+@main
+struct iSpendLiveActivityBundle: WidgetBundle {
+    var body: some Widget { SpendingLiveActivityWidget() }
+}